Module caesar_core::sender::client

Structs

Context ğŸ”’

The context of a sender.

File ğŸ”’

Struct representing a file to be sent.

Constants

DELAY ğŸ”’

DESTINATION ğŸ”’

MAX_CHUNK_SIZE ğŸ”’

NONCE_SIZE ğŸ”’

Functions

on_chunk ğŸ”’

Asynchronously transfers the chunks of files to the receiver.

on_create_room ğŸ”’

Handles the create room packet.

on_error ğŸ”’

Handles errors by returning a Status with the error message.

on_handshake ğŸ”’

Handles the handshake response packet received from the receiver.

on_handshake_finalize ğŸ”’

Finalizes the handshake by sending the list of files to the receiver and starting the file transfer task.

on_join_room ğŸ”’

Handle the join room packet.

on_leave_room ğŸ”’

Handle the leave room packet.

on_message ğŸ”’

Handles the incoming message from the WebSocket.

on_progress ğŸ”’

Handle the progress packet.

start

Starts the sender process.